+++ package.json
@@ -32,8 +32,6 @@
     "@types/react": "^18.3.1",
     "@types/react-dom": "^18.3.5",
     "@types/webextension-polyfill": "^0.10.7",
-    "@typescript-eslint/eslint-plugin": "^7.8.0",
-    "@typescript-eslint/parser": "^7.8.0",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.19",
     "chrome-types": "^0.1.320",
