+++ src/pages/content/yt-watch/contentStyle.css
@@ -13,16 +13,15 @@
 .yt-watch-holodex-icon path:nth-child(1) {
   fill: var(--yt-button-icon-button-text-color, var(--yt-spec-text-secondary));
 }
-
 .yt-watch-holodex-icon path:nth-child(2) {
   stroke: var(--yt-button-icon-button-text-color, var(--yt-spec-text-secondary));
 }
 
-.yt-watch-holodex-btn:hover .yt-watch-holodex-icon>path:nth-child(1) {
+.yt-watch-holodex-btn:hover .yt-watch-holodex-icon > path:nth-child(1) {
   fill: #f06292;
 }
 
-.yt-watch-holodex-btn:hover .yt-watch-holodex-icon>path:nth-child(2) {
+.yt-watch-holodex-btn:hover .yt-watch-holodex-icon > path:nth-child(2) {
   stroke: #5da2f2;
 }
 
@@ -42,8 +41,8 @@ yt-holodex-button-shape {
   inset: auto 68px auto auto !important;
 }
 
-@media screen and (max-width: 1200px) {
-  .yt-watch-holodex-label {
-    display: none;
-  }
-}
+/*@media screen and (max-width: 1200px) {*/
+/*  .yt-watch-holodex-label {*/
+/*    display: none;*/
+/*  }*/
+/*}*/
